feat:#46 리뷰 섹션 개선 및 리뷰 텍스트 상수화
