+++ src/constants/reviewConstants.ts
@@ -1,0 +1,9 @@
+/**
+ * Review 항목을 정의한 enum
+ */
+export const REVIEW_TEXTS = [
+  '친절해요',
+  '진료 대기가 없어요',
+  '시설이 좋고 청결해요',
+  '전문적이에요',
+];
